Create Number Guessing Game code in JAVA
In this java program, a random number ranging between 1-100 will be generated and the user needs to guess the number in fixed number of guesses. Hints will be provided to the user after every wrong answer to make it easy for the user to find the number. It's a fun game to play when bored.